Feat: course 동기화 코드 훅으로 분리
